chore: rename log.h and logger.h to a more suitable, verbose name
